add get recommended course api, call it in viewmodel for internal testing

+++ app/src/main/java/com/projectprovip/h1eu/giasu/data/course/dto/learning_course_detail_dto/Value.kt
@@ -10,7 +10,7 @@ data class Value(
     val id: String,
     val lastModificationTime: String,
     val learningMode: String,
-    val sectionFee: Int,
+    val sectionFee: Double,
     val sessionDuration: Int,
     val sessionPerWeek: Int,
     val status: String,
